Match the regexp `^// Code generated .* DO NOT EDIT\.$`
If the generated code does not match the above regexp, tools such as `go
lint` will consider the bindata generated files as source code and give
lint warnings.
As detailed by https://golang.org/s/generatedcode (which links to this
comment by @robpike:
https://github.com/golang/go/issues/13560#issuecomment-288457920), this
is the standard convention for marking generated code in golang.